+++ src/overtone/studio/core.clj
@@ -4,11 +4,15 @@
   (:use [overtone.sc defaults server])
   (:require [overtone.config.log :as log]))
 
-(defonce studio* (atom {:synth-group       nil
-                        :instruments       {}
-                        :instrument-group  nil
-                        :master-volume     DEFAULT-MASTER-VOLUME
-                        :input-gain        DEFAULT-MASTER-GAIN
-                        :bus-mixers        {:in []
-                                            :out []}
-                        :recorder          nil}))
+(defonce studio* (atom {:synth-group      nil
+                        :instruments      {}
+                        :instrument-group nil
+                        :master-volume    DEFAULT-MASTER-VOLUME
+                        :input-gain       DEFAULT-MASTER-GAIN
+                        :bus-mixers       {:in  []
+                                           :out []}
+                        :recorder         nil
+                        :aux              {}}))
+
+(comment
+  @studio*)