+++ src/overtone/studio/mixer.clj
@@ -241,18 +241,15 @@
 (defn add-instrument
   "Add an instrument to the session."
   [inst]
-  (let [i-name (:name inst)]
+  (let [i-name (:full-name inst)]
     (swap! studio* assoc-in [:instruments i-name] inst)
     i-name))
 
 (defn remove-instrument
   "Remove an instrument from the session."
-  [i-name]
-  (swap! studio* (fn [s]
-                   (let [insts (:instruments s)
-                         insts (dissoc insts (name i-name))]
-                     (assoc s :instruments insts))))
-  (event :inst-removed :inst-name i-name))
+  [full-name]
+  (swap! studio* update :instruments dissoc full-name)
+  (event :inst-removed :inst-name full-name))
 
 (defn clear-instruments
   "Clear all instruments from the session."
@@ -260,8 +257,4 @@
   (let [[{:keys [instruments]} _]
         (swap-vals! studio* assoc :instruments {})]
     (doseq [[_name inst] instruments]
-      (group-free (:instance-group inst)))))
-
-(comment
-  (group-clear (:instrument-group @studio*))
-  (pp-node-tree))
+      (group-free (:group inst)))))