+++ CHANGELOG.md
@@ -74,6 +74,7 @@ reconcile the two.
 ## Changed
 
 - [at-at] Bumped at-at to 1.4.65, which fixes reflection warnings.
+- [insts] sampled-piano now takes an `:amp` arg, as per synth/inst conventions
 
 # 0.14.3199 (2024-05-19 / 5d1c1ed)